+++ includes/footer.php
@@ -1,0 +1,3 @@
+</body>
+
+</html>
