+++ includes/variables.php
@@ -1,0 +1,9 @@
+
+        <?php
+$myproducts = [
+   ["Bonnet en laine", 10, "€", "","Description : Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a leo diam. Quisque lorem orci, accumsan quis dolor sed, gravida.","bonnet 1.jpeg"] ,
+   ["Bonnet en laine bio", 20, "€","","Description : Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a leo diam. Quisque lorem orci, accumsan quis dolor sed, gravida.","bonnet 2.jpeg"],
+  ["Bonnet en laine et cachemire", 12, "€","","Description : Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a leo diam. Quisque lorem orci, accumsan quis dolor sed, gravida.","bonnet 3.jpeg"],
+  ["Bonnet arc-en-ciel", 15, "€","","Description : Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis a leo diam. Quisque lorem orci, accumsan quis dolor sed, gravida.","bonnet 4.jpeg"]
+]
+?>
